% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_session.R
\name{class_session}
\alias{class_session}
\title{Zoom Class Session Summary}
\usage{
class_session(
  data,
  class_start = NULL,
  class_end = NULL,
  arrange_by = c("Join_Time", "Name", "None"),
  arrange_group = c("Name-Email", "Session", "None"),
  period_to = c("period", "second", "minute", "hour"),
  round_digits = NULL
)
}
\arguments{
\item{data}{A data frame (tibble) with class \code{zoom_participants}}

\item{class_start}{(Character) The time of class started, input as "hh:mm:ss".
If \code{NULL} (default) and \code{zoom_participants} has \code{meeting_overview} attribute, "Start_Time" of the attribute will be used.}

\item{class_end}{(Character) The time of class ended, input as "hh:mm:ss".
If \code{NULL} (default) and \code{zoom_participants} has \code{meeting_overview} attribute, "End_Time" of the attribute will be used.}

\item{arrange_by}{(Character) How to arrange rows, must be one of:
\itemize{
\item \strong{"Join_Time"}: arrange by \code{Join_Time}, earliest first
\item \strong{"Name"}: arrange by \verb{Name (Original Name)}
\item \strong{"None"}: no arrange, preserve rows as original Zoom's participant file
}}

\item{arrange_group}{(Character) Indicate group of \code{arrange_by} will be affected.
\itemize{
\item \strong{"Name-Email"}: group by \verb{Name (Original Name)} and \code{Email}
\item \strong{"Session"}: group by \code{Session}
\item \strong{"None"}: no grouping variable
}}

\item{period_to}{(Character) Indicate the units of "Before_Class", "During_Class", "After_Class" and "Total_Time" in the output.
Must be one of:
\itemize{
\item \strong{"period"} (default): return as \code{lubridate::period} object
\item \strong{"second"}: return as seconds (numeric)
\item \strong{"minute"}: return as minute (numeric)
\item \strong{"hour"}: return as hour (numeric)
}}

\item{round_digits}{(Integer) significant digits to round the numeric columns. If \code{NULL} (default), no round.}
}
\value{
A data frame (tibble) with class \code{zoom_class}. It has the following columns:
\itemize{
\item \strong{"Name (Original Name)"}: the same column "Name (Original Name)" as input file
\item \strong{"Name"}: Current name displayed in Zoom meeting of the participants.
\item \strong{"Name_Original"}: Original name of the participants.
It was extracted from the contents within the last balanced parentheses of "Name (Original Name)".
\item \strong{"Email"}: from original "User Email" column
\item \strong{"Session"}: Indicate active session(s) in Zoom of each students.
Computed by ranking "Join_Time" in the grouping variables: "Name (Original Name)" and "Email".
\item \strong{"Class_Start"} (POSIXct): Compute from date of "Join_Time" with time specified by \code{class_start} argument.
\item \strong{"Class_End"} (POSIXct): Compute from date of "Leave_Time" with time specified by \code{class_end} argument.
\item \strong{"Join_Time"}: from original "Join Time" column
\item \strong{"Leave_Time"}: from original "Leave Time" column
\item \strong{"Before_Class"}: Time spent before \code{class_start} of each session.
\item \strong{"During_Class"}: Time spent during class (between \code{class_start} and \code{class_end}) of each session.
\item \strong{"After_Class"}: Time spent after \code{class_end} of each session.
\item \strong{"Total_Time"}: "Before_Class" + "During_Class" + "After_Class"
\item \strong{"Duration_Minutes"}: from original "Duration (Minutes)" column
\item \strong{"Rec_Consent"}: from original "Recording Consent" column
\item \strong{"Multi_Device"}: \code{TRUE} if any sessions of each student joined Zoom with multiple devices.
It is computed by checking whether "Join_Time" and "Leave_Time" was overlapped with other sessions within each students.
Otherwise, \code{NA} will be returned because there is no way to be certain that students joined with a single device.
}
}
\description{
\code{\link[=class_session]{class_session()}} was designed to retrieve time information about each student's session
in Zoom's classroom at given time period. Input of this function (\code{data}) is a \code{zoom_participants} object as
returned by \code{\link[=read_participants]{read_participants()}}.
}
\details{
By providing time of class started (\code{class_start}) and ended (\code{class_end}),
this function will compute time spent before, during, and after class of each sessions.
If each student joined the classroom using multiple device, \code{TRUE} will be placed in the "Multi_Device"
column of the output.

The output will displayed one row per session of each student.
(e.g., If student "A" joined more than one time, there will be multiple rows of "A").
Finally, rows will be arranged as specified by \code{arrange_by} with grouping by \code{arrange_group}.
}
\seealso{
Checkout \code{\link[=class_students]{class_students()}} for summary of each students.
}
