% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_studentsID.R
\name{class_studentsID}
\alias{class_studentsID}
\title{Zoom Class Students ID Summary}
\usage{
class_studentsID(
  data,
  id_regex = ".*",
  collapse = "; ",
  class_start = NULL,
  class_end = NULL,
  late_cutoff = NULL,
  period_to = c("period", "second", "minute", "hour"),
  round_digits = NULL
)
}
\arguments{
\item{data}{A data frame (tibble) with class \code{zoom_participants}}

\item{id_regex}{(Character) regular expression used to extract student's ID from \verb{Name (Original Name)} to \code{ID} column}

\item{collapse}{(Character) How to collapse "Name (Original Name)" and "Email" if found multiple different names or Email of each student's ID.}

\item{class_start}{(Character) The time of class started, input as "hh:mm:ss" or "hh:mm".
If \code{NULL} (default) and \code{zoom_participants} has \code{meeting_overview} attribute, "Start_Time" of the attribute will be used.}

\item{class_end}{(Character) The time of class ended, input as "hh:mm:ss" or "hh:mm".
If \code{NULL} (default) and \code{zoom_participants} has \code{meeting_overview} attribute, "End_Time" of the attribute will be used.}

\item{late_cutoff}{(Character) Late time cutoff (input as "hh:mm:ss" or "hh:mm"). If provided, "Late_Time" will be included in the output columns.}

\item{period_to}{(Character) Indicate the units of "Before_Class", "During_Class", "After_Class" and "Total_Time" in the output.
Must be one of:
\itemize{
\item \strong{"period"} (default): return as \code{lubridate::period} object
\item \strong{"second"}: return as seconds (numeric)
\item \strong{"minute"}: return as minute (numeric)
\item \strong{"hour"}: return as hour (numeric)
}}

\item{round_digits}{(Integer) significant digits to round the numeric columns. If \code{NULL} (default), no round.}
}
\value{
A data frame (tibble) with class \code{zoom_class}. It has the following columns:
\itemize{
\item \strong{"ID"}: the student's ID as extracted from \verb{Name (Original Name)} by \code{id_regex}. Unmatched ID will leave as \code{NA} in the bottom rows.
\item \strong{"Name"}: Name combinations from \verb{Name (Original Name)} of each IDs
\item \strong{"Email"}: Email combinations from \code{Email} of each IDs
\item \strong{"Session_Count"}: Show counts of how many session that each students joined or leaved Zoom class.
\item \strong{"Class_Start"} (POSIXct): Compute from date of "Join_Time" with time specified by \code{class_start} argument.
\item \strong{"Class_End"} (POSIXct): Compute from date of "Leave_Time" with time specified by \code{class_end} argument.
\item \strong{"First_Join_Time"}: First join time of each students
\item \strong{"Last_Leave_Time"}: Last leave time of each students
\item \strong{"Before_Class"}: Time spent before \code{class_start} of each student.
\item \strong{"During_Class"}: Time spent during class (between \code{class_start} and \code{class_end}) of each student.
\item \strong{"After_Class"}: Time spent after \code{class_end} of each student.
\item \strong{"Total_Time"}: "Before_Class" + "During_Class" + "After_Class"
\item \strong{"Duration_Minutes"}: Sum of "Duration (Minutes)" for each students.
\strong{Notice:} \code{Total_Time} is likely to be less than original \code{Duration_Minutes} because the latter round the decimal down.
This difference will be more pronounced when summed with multiple sessions.
\item \strong{"Multi_Device"}: \code{TRUE} if students joined Zoom with multiple devices in any session.
\item \strong{"Late_Time"} (Optional): If provide \code{late_cutoff} as "hh:mm:ss", "Late_Time" period is computed by \code{Join_Time} - \code{late_cutoff}.
}
}
\description{
Similar to \code{\link[=class_students]{class_students()}}, \code{\link[=class_studentsID]{class_studentsID()}} was designed to retrieve time information about each students
in Zoom's classroom at given time period. Input of this function (\code{data}) is a \code{zoom_participants} object as
returned by \code{\link[=read_participants]{read_participants()}}.
}
\details{
Firstly, student's ID will be extracted from "Name (Original Name)" column using regular expression provided by
\code{id_regex} into a new "ID" column. The "ID" column will be used as a groping variable that represents each students.
"Name" and "Email" column in the output is a combination (in each IDs) of "Name (Original Name)" and "Email", respectively.

By providing time of class started (\code{class_start}) and ended (\code{class_end}),
this function will compute time spent before, during, and after class of each students (\code{ID}).
Furthermore, Late time period can be calculated from students who joined Zoom class later than certain cutoff time (\code{late_cutoff}).
If each students joined the classroom using multiple device in any sessions, \code{TRUE} will be placed in the "Multi_Device"
column of the output.

The result will be displayed one row per student's ID.
"Session_Count" will counts how many times each students join or leave the Zoom class.
}
\seealso{
Checkout \code{\link[=class_students]{class_students()}} for summary by each student, and \code{\link[=class_session]{class_session()}} for summary of each session.
}
