% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{create_zoom_class}
\alias{create_zoom_class}
\title{Create zoom_class}
\usage{
create_zoom_class(x, class_start, class_end, late_cutoff = NA)
}
\arguments{
\item{x}{a "zoom_participants" object}

\item{class_start}{Vector included in "class_overview" attribute.}

\item{class_end}{Vector included in "class_overview" attribute.}

\item{late_cutoff}{Vector included in "class_overview" attribute.}
}
\value{
a "zoom_class" object with "class_overview" attribute.
}
\description{
Create "zoom_class" object with "class_overview" attribute.
}
