% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{create_zoom_participants}
\alias{create_zoom_participants}
\title{Create zoom_participants}
\usage{
create_zoom_participants(x, meeting_overview)
}
\arguments{
\item{x}{object}

\item{meeting_overview}{attribute to add}
}
\value{
a "zoom_participants" object with "meeting_overview" attribute
}
\description{
Create "zoom_participants" class and add "meeting_overview" attribute.
}
