% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_participants.R
\name{extract_CurrOrig_name}
\alias{extract_CurrOrig_name}
\title{Extract Current and Original name to a Data.frame}
\usage{
extract_CurrOrig_name(string)
}
\arguments{
\item{string}{Character vector}
}
\value{
A tibble with \code{current} and \code{original} column names
}
\description{
Extract Current and Original name to a Data.frame
}
