% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-time.R
\name{get_after_class_time}
\alias{get_after_class_time}
\title{Get time after class}
\usage{
get_after_class_time(end, join, leave)
}
\arguments{
\item{end}{(POSIXct) class end time}

\item{join}{(POSIXct) class join time}

\item{leave}{(POSIXct) class leave time}
}
\value{
Period Vector
}
\description{
Get time after class as a period objects.
Compute time period from class end time or joined time, depending on which comes later, to
leave time.
}
