% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-time.R
\name{get_before_class_time}
\alias{get_before_class_time}
\title{Get time before class}
\usage{
get_before_class_time(start, join, leave)
}
\arguments{
\item{start}{(POSIXct) class start time}

\item{join}{(POSIXct) join time}

\item{leave}{(POSIXct) leave time}
}
\value{
Period vector
}
\description{
Get time before class as a period objects.
Compute from time after joined class to class start time or leave time, depending
on which comes earlier.
}
