% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-time.R
\name{get_during_class_time}
\alias{get_during_class_time}
\title{Get time during class}
\usage{
get_during_class_time(start, end, join, leave)
}
\arguments{
\item{start}{(POSIXct) Class start time}

\item{end}{(POSIXct) Class end time}

\item{join}{(POSIXct) Class join time}

\item{leave}{(POSIXct) Class leave time}
}
\value{
Period Vector
}
\description{
Get time during class as a period objects.
Compute time period after class start or joined time (choose the later time) to class end time
or class leave time (choose the earlier time).
}
