% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-sum.R
\name{get_late_time}
\alias{get_late_time}
\title{Get Late Time}
\usage{
get_late_time(join, late_cutoff)
}
\arguments{
\item{join}{(POSIXct) Join Time}

\item{late_cutoff}{(POSIXct) Cutoff time, after this time will be considered late.}
}
\value{
Period vector
}
\description{
Get late time if joined after \code{late_cutoff}, otherwise \code{NA} is returned.
}
