% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoom_chat.R
\name{is_chat_full}
\alias{is_chat_full}
\title{Test if Zoom Chat is Full}
\usage{
is_chat_full(chr)
}
\arguments{
\item{chr}{Character vector to test}
}
\value{
Logical: \code{TRUE} if it was Full
}
\description{
Full Zoom chat has \verb{hh:mm:ss from <Name> to <Target> :} structure.
}
