% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-session-device.R
\name{is_overlap}
\alias{is_overlap}
\title{Is interval overlapping?}
\usage{
is_overlap(lwr, upr)
}
\arguments{
\item{lwr}{(Numeric vector) Lower bound of the interval}

\item{upr}{(Numeric vector) Upper bound of the interval}
}
\value{
Logical vector corresponding to each elements of numeric vector
}
\description{
Pairwise testing for each interval whether it is overlap with any others (exclusively).
}
