% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{is_zoom_participants}
\alias{is_zoom_participants}
\title{Is zoom_participants}
\usage{
is_zoom_participants(x)
}
\arguments{
\item{x}{object}
}
\value{
Logical
}
\description{
Check whether an object inherits "zoom_participants" class.
}
