% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-sum.R
\name{late_cutoff_POSIXct}
\alias{late_cutoff_POSIXct}
\title{Get Late Cutoff as POSIXct}
\usage{
late_cutoff_POSIXct(join_POSIXct, late_cutoff = NULL)
}
\arguments{
\item{join_POSIXct}{(POSIXct) Join time}

\item{late_cutoff}{(Character) Late Cutoff as "hh:mm:ss"}
}
\value{
\code{late_cutoff} as POSIXct
}
\description{
Get Late Cutoff as POSIXct
}
