% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-attr.R
\name{meeting_overview}
\alias{meeting_overview}
\title{Meeting Overview}
\usage{
meeting_overview(data)
}
\arguments{
\item{data}{"zoom_participants" object}
}
\value{
"meeting_overview" attributes:
\itemize{
\item A tibble with following column names: "Meeting_ID", "Topic", "Start_Time", "End_Time", "Email", "Duration_Minute", "Participants"
}
}
\description{
Retrieve overview of meeting information such as "Meeting_ID", "Topic", etc. from "zoom_participants" object.
}
