% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{new_zoom_class}
\alias{new_zoom_class}
\title{New zoom_class}
\usage{
new_zoom_class(x)
}
\arguments{
\item{x}{"zoom_participants" object}
}
\value{
"zoom_class" object
}
\description{
Add "zoom_class" class to "zoom_participants" object
}
