% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{new_zoom_participants}
\alias{new_zoom_participants}
\title{New zoom_participants}
\usage{
new_zoom_participants(x = data.frame())
}
\arguments{
\item{x}{object}
}
\value{
a "zoom_participants" object
}
\description{
Construct "zoom_participants" class (must be data.frame)
}
