% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_participants.R
\name{parse_pp_datetime}
\alias{parse_pp_datetime}
\title{Parse Date-Time formatting Using Lubridate Function}
\usage{
parse_pp_datetime(chr)
}
\arguments{
\item{chr}{Character vector of date-time}
}
\value{
A POSIXct
}
\description{
Parse Date-Time formatting Using Lubridate Function
}
