% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-time.R
\name{parse_time_flex}
\alias{parse_time_flex}
\title{Flexible Parse Time}
\usage{
parse_time_flex(chr)
}
\arguments{
\item{chr}{Input can be "hh:mm:ss" or "hh:mm"}
}
\value{
A period object
}
\description{
Flexible Parse Time
}
