% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-stu-id.R
\name{paste_unique_collapse_na.rm}
\alias{paste_unique_collapse_na.rm}
\title{Paste Collapse with with Unique value & NA removed}
\usage{
paste_unique_collapse_na.rm(x, collapse = NULL)
}
\arguments{
\item{x}{A vector to paste and collapse}

\item{collapse}{an optional character string to separate the results.}
}
\value{
A character vector
}
\description{
Paste Collapse with with Unique value & NA removed
}
