% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-helpers.R
\name{period_to_dbl}
\alias{period_to_dbl}
\title{Convert Period Object to Double}
\usage{
period_to_dbl(x, unit = c("second", "minute", "hour"), round_digits = NULL)
}
\arguments{
\item{x}{period object}

\item{unit}{(Character) specify unit of the converted numeric output}

\item{round_digits}{significant digits to round. If \code{NULL}, no round.}
}
\value{
a double vector
}
\description{
Convert Period Object to Double
}
