% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-arrange.R
\name{process_class_arrange}
\alias{process_class_arrange}
\title{Arrange}
\usage{
process_class_arrange(
  df_cleaned,
  arrange_by = c("Join_Time", "Name", "None"),
  arrange_group = c("Name-Email", "Session", "None")
)
}
\arguments{
\item{df_cleaned}{\code{zoom_participants} tibble}

\item{arrange_by}{(Character) How to arrange rows, must be one of:
\itemize{
\item \strong{"Join_Time"}: arrange by \code{Join_Time}, earliest first
\item \strong{"Name"}: arrange by \verb{Name (Original Name)}
\item \strong{"None"}: no arrange, preserve rows as original Zoom's participant file
}}

\item{arrange_group}{(Character) Indicate group of \code{arrange_by} will be affected.
\itemize{
\item \strong{"Name-Email"}: group by \verb{Name (Original Name)} and \code{Email}
\item \strong{"Session"}: group by \code{Session}
\item \strong{"None"}: no grouping variable
}}
}
\value{
An arranged tibble
}
\description{
Arrange \code{zoom_participants}.
First, group by \code{arrange_group} then perform arrange in \code{arrange_by}.
}
