% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-session-device.R
\name{process_class_session_device}
\alias{process_class_session_device}
\title{Process Class Session & Multiple Device}
\usage{
process_class_session_device(df_cleaned)
}
\arguments{
\item{df_cleaned}{Cleaned data.frame}
}
\value{
A tibble with column "Session" and "Multi_Device" added.
}
\description{
Add column \code{Session} for each grouping (\verb{Name (Original Name)}, \code{Email}) by ranking \code{Join_Time}.
If \code{Join_Time} is exactly the same for each grouping, ranking will not increase.
Add column \code{Multi_Device} that check whether sessions of each grouping are overlapped with any others.
For overlapping sessions, \code{TRUE} will be placed at those sessions, otherwise \code{NA}.
}
