% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-stu-id.R
\name{process_class_studentsID}
\alias{process_class_studentsID}
\title{Process Class Student ID}
\usage{
process_class_studentsID(df_processed, id_regex = ".*", late_cutoff = NULL)
}
\arguments{
\item{df_processed}{time-processed tibble}

\item{id_regex}{(Character) regular expression used to extract student's ID from \verb{Name (Original Name)} to \code{ID} column}

\item{late_cutoff}{(Character) Cutoff time (input as "hh:mm:ss"), participants join after this time will be considered late.}
}
\value{
a summary tibble
}
\description{
Extract student IDs from \verb{Name (Original Name)} column using regular expression (\code{id_regex}) into \code{ID} column.
Then, compute summary of each IDs in terms of:
\code{Name} (a combination of \verb{Name (Original Name)} of each IDs),
\code{Email} (a combination of \code{Email} of each IDs),
\code{Session_Count} (total sessions of each student),
\code{First_Join_Time}, \code{Last_Leave_Time}, \code{Duration_Minutes}, \code{Multi_Device}.
Has an option to check whether student joined class later than specific time cutoff
and the late period will be include in \code{Late_Time} column.
}
