% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-sum.R
\name{process_class_summary}
\alias{process_class_summary}
\title{Process Class Summary}
\usage{
process_class_summary(df_processed, late_cutoff = NULL)
}
\arguments{
\item{df_processed}{time-processed tibble}

\item{late_cutoff}{(Character) Cutoff time (input as "hh:mm:ss"), participants join after this time will be considered late.}
}
\value{
a summary tibble
}
\description{
Compute summary of each student in terms of: \code{Session_Count} (total sessions of each student),
\code{First_Join_Time}, \code{Last_Leave_Time}, \code{Duration_Minutes}, \code{Multi_Device}.
Has an option to check whether student joined class later than specific time cutoff
and the late period will be include in \code{Late_Time} column.
}
