% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-time.R
\name{process_class_time}
\alias{process_class_time}
\title{Process Class Time}
\usage{
process_class_time(df_cleaned, class_start = NULL, class_end = NULL)
}
\arguments{
\item{df_cleaned}{A "zoom_participants" object (cleaned tibble)}

\item{class_start}{(Character) Time of class start (input as "hh:mm:ss"). If \code{NULL}, use "Start_Time" from "meeting_overview" attribute.}

\item{class_end}{(Character) Time of class end (input as "hh:mm:ss"). If \code{NULL}, use "End_Time" from "meeting_overview" attribute.}
}
\value{
A tibble with class "zoom_class" with the following columns added:
\itemize{
\item \strong{"Class_Start"}: POSIXct of \code{class_start}
\item \strong{"Class_End"}: POSIXct of \code{class_end}
\item \strong{"Before_Class"}: Period before class started of each session
\item \strong{"During_Class"}: Period during class of each session
\item \strong{"After_Class"}: Period after class ended of each session
}
}
\description{
By providing \code{class_start} and \code{class_end}, the function compute time period before, during, and after class of
each sessions.
}
