% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-class-stu-id.R
\name{process_summary_ID}
\alias{process_summary_ID}
\title{Process Summary for Extracted ID Data Frame}
\usage{
process_summary_ID(df_extracted, id_type = c("chr", "NA"), collapse = "; ")
}
\arguments{
\item{df_extracted}{Extracted ID data frame}

\item{id_type}{Whether \code{df_extracted} has ID column as character (match) or \code{NA} (no match)}

\item{collapse}{How to collapse "Name (Original Name)" and "Email" column}
}
\value{
A data frame with column: "ID" (Chr or \code{NA}), "Name", "Email", etc.
}
\description{
Process Summary for Extracted ID Data Frame
}
