% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_participants.R
\name{read_meeting_overview}
\alias{read_meeting_overview}
\title{Read Meeting Details from Zoom's Participant File}
\usage{
read_meeting_overview(file)
}
\arguments{
\item{file}{A path to a Zoom's participant file}
}
\value{
a tibble or \code{NULL}
}
\description{
If "Meeting ID" was detected in the column names, read only 1 row (and column names) into a tibble and clean column names.
If not, return \code{NULL}.
}
