% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_participants.R
\name{read_participants}
\alias{read_participants}
\title{Read & Clean Zoom's Participant File}
\usage{
read_participants(file)
}
\arguments{
\item{file}{A path to a Zoom's participant file}
}
\value{
A data frame (tibble) with class "zoom_participants". It has the following new columns:
\itemize{
\item \strong{"Name (Original Name)"}: the same column "Name (Original Name)" as input file
\item \strong{"Name"}: Current name displayed in Zoom meeting of the participants.
\item \strong{"Name_Original"}: Original name of the participants.
It was extracted from the contents within the last balanced parentheses of "Name (Original Name)".
\item \strong{"Email"}: from original "User Email" column
\item \strong{"Join_Time"}: from original "Join Time" column
\item \strong{"Leave_Time"}: from original "Leave Time" column
\item \strong{"Duration_Minutes"}: from original "Duration (Minutes)" column
\item \strong{"Rec_Consent"}: from original "Recording Consent" column
}
}
\description{
This function read Zoom's participant file from CSV to a tibble with sensible formatting and cleaner
column names. \code{readr::read_csv} will be used to read the data. \verb{Join Time} and \verb{Leave Time} will be formatted as POSIXct object.
Two new columns, \code{Name} and \code{Name_Original}, will be extracted from \verb{Name (Original Name)}.
Every column names will be cleaned so that it easier to type in \code{R}.
If Zoom's participant file has meeting details such as "Meeting ID", "Topic", etc. at the first row,
it will be assign to "meeting_overview" attributes.
}
