% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoom_chat.R
\name{read_zoom_chat}
\alias{read_zoom_chat}
\title{Read Zoom Chat File}
\usage{
read_zoom_chat(file, format_time = FALSE)
}
\arguments{
\item{file}{Path to Zoom's chat .txt file(s)}

\item{format_time}{(Logical) if \code{TRUE} column "Time" will be formatted to period object}
}
\value{
A tibble
}
\description{
Read Zoom chat from file path into a tibble.
}
