% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readzoom-package.R
\docType{package}
\name{readzoom-package}
\alias{readzoom}
\alias{readzoom-package}
\title{readzoom: Read Zoom}
\description{
This package contains a collection of functions for parsing, data wrangling and analysis of program Zoom chat file & participant recording.
}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
