% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-helpers.R
\name{vars_period_to_dbl}
\alias{vars_period_to_dbl}
\title{Convert Period Variable in DF to Double}
\usage{
vars_period_to_dbl(
  data,
  unit = c("second", "minute", "hour"),
  round_digits = NULL
)
}
\arguments{
\item{data}{a data.frame}

\item{unit}{(Character) specify unit of the converted numeric output}

\item{round_digits}{significant digits to round. If \code{NULL}, no round.}
}
\value{
A data.frame with period object vars converted to numeric with specified unit.
}
\description{
Suffix \verb{*_unit} will be added to the converted column names.
}
