% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_chat_count.R
\name{zoom_chat_count}
\alias{zoom_chat_count}
\title{Count Zoom Chat Message}
\usage{
zoom_chat_count(
  data,
  collapse = " ~ ",
  extract_id = FALSE,
  id_regex = "[:digit:]+"
)
}
\arguments{
\item{data}{a tibble with class "zoom_chat".}

\item{collapse}{(Character) deliminator character to separate each entries of participants}

\item{extract_id}{(Character) Whether to extract ID number from \code{Name} column or not.}

\item{id_regex}{(Character) If \code{extract_id = TRUE}, This would specify regular expression (\code{{stringr}}'s style) to extract ID.}
}
\value{
A tibble with the following columns:
\itemize{
\item \strong{ID}: If \code{extract_id = TRUE}, this is the ID of the participants.
\item \strong{Name}: Name of each participants
\item \strong{Message_Count}: Counts of how many times each participants replies
\item \strong{Content}: All replies of each participants, separated by \code{collapse} argument.
}
}
\description{
Count how many times each participants replies in the Zoom chat box and also group messages entries per participants.
}
