% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoom_chat.R
\name{zoom_chat_ext_contents}
\alias{zoom_chat_ext_contents}
\title{Extract Contents from Zoom Chat}
\usage{
zoom_chat_ext_contents(chr)
}
\arguments{
\item{chr}{A character vector}
}
\value{
A character vector
}
\description{
Extract Contents from Zoom Chat
}
