% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoom_chat.R
\name{zoom_chat_ext_contents_abbr}
\alias{zoom_chat_ext_contents_abbr}
\title{Extract Contents from Zoom Chat Abbreviated}
\usage{
zoom_chat_ext_contents_abbr(chr)
}
\arguments{
\item{chr}{A character vector}
}
\value{
A character vector
}
\description{
Extract Contents from Zoom Chat Abbreviated
}
