% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoom_chat.R
\name{zoom_chat_ext_name_target}
\alias{zoom_chat_ext_name_target}
\title{Extract <Name> and <Target>}
\usage{
zoom_chat_ext_name_target(chr)
}
\arguments{
\item{chr}{A character vector}
}
\value{
A data.frame containing "Name" and "Target"
}
\description{
Extract <Name> and <Target>
}
