% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoom_chat.R
\name{zoom_chat_ext_time}
\alias{zoom_chat_ext_time}
\title{Extract <Time> Stamp}
\usage{
zoom_chat_ext_time(chr)
}
\arguments{
\item{chr}{A character vector}
}
\value{
character vector
}
\description{
Extract <Time> Stamp
}
