% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zoom_chat.R
\name{zoom_chat_extract}
\alias{zoom_chat_extract}
\title{Extract Each Element to Tibble}
\usage{
zoom_chat_extract(x, format_time = FALSE)
}
\arguments{
\item{x}{A character vector}

\item{format_time}{(Logical) if \code{TRUE} column "Time" will be formatted to period object}
}
\value{
A "zoom_chat" subclass Tibble with columns "Time", "Name", "Content".
If full Zoom Chat format is recognized "Target" is added.
}
\description{
Parse Zoom chat as a character vector to a tibble.
}
