% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomclass-package.R
\docType{package}
\name{zoomclass-package}
\alias{zoomclass}
\alias{zoomclass-package}
\title{zoomclass: Zoom Class}
\description{
This package contains a collection of functions for analysis of Zoom classroom report. This includes parsing, data wrangling and analysis of program Zoom chat file & participant recording with option to compute time interval before, during, and after class and also late time.
}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
