% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{zoomclass_example}
\alias{zoomclass_example}
\title{Get path to raw zoomclass example}
\usage{
zoomclass_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
zoomclass comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
library(zoomclass)
zoomclass_example()
zoomclass_example("participants_heroes.csv")
}
