% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-attr.R
\name{class_overview}
\alias{class_overview}
\title{Class Overview}
\usage{
class_overview(data)
}
\arguments{
\item{data}{"zoom_class" object}
}
\value{
"class_overview" attributes:
\itemize{
\item A tibble with following column names: "class_start", "class_end", "late_cutoff"
}
}
\description{
Class Overview
}
