% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example.R
\docType{data}
\name{heroes_students}
\alias{heroes_students}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 72 rows and 3 variables:
\describe{
\item{ID}{Student's ID}
\item{Name}{Student's Name}
\item{Comment}{Character, Provide descriptions about students.}
}
}
\usage{
heroes_students
}
\description{
A simulated data set containing names and student's ID of heroes who enrolled in a semester.
}
\keyword{datasets}
