% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomclass-package.R
\docType{package}
\name{zoomclass-package}
\alias{zoomclass}
\alias{zoomclass-package}
\title{zoomclass: Zoom Class}
\description{
Data analysis of Zoom classroom report. This includes parsing, data wrangling and analysis of program Zoom chat file & participant recording with option to compute time interval before, during, after class, and late time of students.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lightbridge-ks.github.io/zoomclass}
}

}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
